Add extendability to the FocusCropService
With this change we move the file name geration to a simple method.
This allows the user to at least xclass the FocusCropService and
implement another name geration logic.

To make it even easier to hook into the name generation I also added a
singal in front of the old logic so that the user does not need to
xclass the whole serivce.