Return FileRefernce object from getViewHelperImage
If we use the same image with different focus point settings twice we
lose the settings for the first one. This is because the
getViewHelperImage method does not return the FileReference but the
originalFile object.

I don't know why this is the case but my guess would be that it is an
error? We can however return the FileReference object instead of the
File object because they both implement the FileInterface.

Please have a look at the changes, thank you!

Fixes lochmueller/focuspoint#29